backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
yfinance>=1.2.0
pandas==2.2.3
numpy==1.26.4
requests==2.32.3
urllib3==2.2.3
google-generativeai==0.8.3
python-multipart==0.0.20
python-dotenv==1.0.1
